requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
datadog==0.44.0
defusedxml==0.7.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
jira==3.5.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sqlite-vss==0.1.2
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
urllib3==1.26.12
